+++ src/frontend/components/DashboardVideoPaneDownloadOption/index.spec.tsx
@@ -46,6 +46,10 @@ describe('<DashboardVideoPaneDownloadOption />', () => {
       },
     },
     should_use_subtitle_as_transcript: false,
+    playlist: {
+      title: 'foo',
+      lti_id: 'foo+context_id',
+    },
   };
 
   it('renders with checkbox not checked', () => {