+++ CHANGELOG.md
@@ -16,7 +16,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Use `UploadableFileMixin` on `AbstractImage` model
-- Fix how permissions are used to display dashboard
+- Fix dashboard read versus update permissions in situations 
+  of playlist portability
 
 ## [3.9.1] - 2020-06-24